requirements.txt
datasets
huggingface_hub
numpy
pandas
sentence-transformers
torch
transformers
tqdm
scipy
unbabel-comet
